fixing md5 checks

+++ download.py
@@ -3,7 +3,7 @@ import hashlib
 import json
 import os
 import os.path as osp
-
+import warnings
 import requests
 from jsonargparse import CLI
 from tqdm import tqdm
@@ -11,9 +11,11 @@ from tqdm.contrib.concurrent import thread_map
 
 
 def download_item(url: str, dst: str, chunk_size: int = 1024 * 1024):
+    if osp.exists(dst):
+        return
+
     response = requests.get(url, stream=True)
     filename = osp.basename(dst)
-    check_md5 = hashlib.md5()
 
     with open(dst, mode="wb") as file:
         with tqdm(
@@ -26,8 +28,14 @@ def download_item(url: str, dst: str, chunk_size: int = 1024 * 1024):
         ) as pbar:
             for chunk in response.iter_content(chunk_size=chunk_size):
                 file.write(chunk)
-                check_md5.update(chunk)
                 pbar.update(len(chunk))
+
+
+def md5(file_path: str) -> int:
+    check_md5 = hashlib.md5()
+
+    with open(file_path, mode="rb") as f:
+        check_md5.update(f.read())
 
     return check_md5.hexdigest()
 
@@ -47,23 +55,18 @@ def download(root: str, max_threads: int = 8):
     folders = ["", "validation", *["training"] * (len(file_list) - 2)]
     dst_files = [osp.join(dst_root, d, f["name"]) for f, d in zip(file_list, folders)]
 
-    os.makedirs(osp.join(root, "validation"), exist_ok=True)
-    os.makedirs(osp.join(root, "training"), exist_ok=True)
+    os.makedirs(osp.join(dst_root, "validation"), exist_ok=True)
+    os.makedirs(osp.join(dst_root, "training"), exist_ok=True)
     urls = [f["download_url"] for f in file_list]
-    actual_md5 = list(
-        thread_map(download_item, urls, dst_files, max_workers=max_threads)
-    )
-    expect_md5 = [f["computed_md5"] for f in file_list]
+    list(thread_map(download_item, urls, dst_files, max_workers=max_threads))
+    actual_md5 = list(thread_map(md5, dst_files, max_workers=max_threads))
+    expect_md5 = set([f["computed_md5"] for f in file_list])
 
-    if not all(a == b for a, b in zip(actual_md5, expect_md5)):
-        raise RuntimeError(
-            (
-                "Downloaded version differs from expected checksum.\n"
-                "Please retry this download and raise an issue in\n\n"
-                "https://github.com/graphcore-research/qm1b-dataset/issues."
-            )
-        )
+    for actual, f in zip(actual_md5, dst_files):
+        if actual not in expect_md5:
+            warnings.warn(f"Unexpected MD5 digest for file: {f}")
 
+    print(f"QM1B dataset downloaded to: {dst_root}")
 
 if __name__ == "__main__":
     CLI(download)
